+++ src/main/java/com/avicagan/bloodandbones/BloodAndBones.java
@@ -58,6 +58,7 @@ public class BloodAndBones {
         modEventBus.addListener(BBBlockEntities::registerCapabilities);
         com.avicagan.bloodandbones.registry.BBDataComponents.COMPONENTS.register(modEventBus);
         com.avicagan.bloodandbones.registry.BBParticles.PARTICLES.register(modEventBus);
+        com.avicagan.bloodandbones.registry.BBItemAttributes.register(modEventBus);
 
         LOGGER.info("Create: Blood & Bones loaded");
     }